Started with parsing the bit contents of the NALU
