+++ lib/util.js
@@ -15,6 +15,23 @@ const util = {
       buf[i] = view[i];
     }
     return hexy.hexy(buf);
+  },
+  readBits: function(byte, numBits) {
+    let array = [];
+    const bitMask = [
+      0x80, 0x40, 0x20, 0x10, 0x08, 0x04, 0x02, 0x01,
+    ];
+    const shift = [
+      0x80, 0x40, 0x20, 0x10, 0x08, 0x04, 0x02, 0x01,
+    ];
+    for(let i = 0; i< numBits; i++) {
+      array.push((byte & bitMask[i]) / shift[i]);
+    }
+    return array;
+  },
+  readFlags: function(byte, numFlags) {
+    const array = this.readBits(byte, numFlags);
+    return array.map(f => f === 1);
   }
 };
